*: implement distinct clause This PR implements the DISTINCT SQL Node. Distinct wraps a projection and emits only rows that have not been previously seen before.
There is a common table with more than 100 rows to perform benchmarks. A Benchmark has been added for both Project and Distinct, which shows that Distinct is two orders of magnitude slower than Project. Although this is just a first approximation to have the feature, in the feature we should spend some time thinking on how to properly and efficiently implement the distinct and/or try to identify if the projection is sorted, which would eliminate the need to keep track of the hashes and make this way faster.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>