sql/analyzer: resolve GetField on schemas with repeated column names Fixes #88
Because of the way we access the fields in the rows using GetField with the index in which they appear and only by name, there was no way of knowing which field was the correct when there were multiple fields with the same name.
Now, all columns include a "Source", which is defined by the table and must match the table name (this is validated by an analyzer validation rule), so that we can identify from which table we should get the field.
One notable exception to this is TableAlias, which avoids this limitation only when its child node is not a Projection (that is, a subquery).

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>