sql/(plan,expression,parse): trace node execution
Fixes #137

This commit adds tracing of spans of execution in all nodes and expressions.

Since expressions and functions evaluation is immediate, spans are finished inside the `Eval` function with a defer.
In the case of plan nodes it's a bit more complex, as they just return an iterator that will be lazily evaluated. Measuring the span as just the time it takes to return the iterator would be wrong, so what it does is wrap all iterators with a `spanIter`, which iterates over an internal iterator and finishes the span when it's needed (there is an error, io.EOF is returned or `Close` is called) pushing some metrics like the number of processed rows or the error without needing the user to provide them.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>