sql/analyzer: replace column indexes with GetFields in SortFields
Fixes #144

Even if literals were being parsed correctly as the Column in the SortFields, we were not acting accordingly on the Sort node, because it acted as a literal and not a GetField expression.
This commit adds a rule to replace the Literal expression inside the SortField Column field into an UnresolvedColumn that will get resolved as any other column in the tree. By doing this with a rule, we avoid handling this special case inside the Sort node.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>