sql: implement functional options pattern for context building
As the number of dependencies inside sql.Context grows, so will the number of arguments of sql.NewContext. To avoid that, since most of them in testing and mocking are the default arguments it's more convenient to set some default values and configure them optionally using the functional options pattern to do so.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>