sql/analyzer: resolve DUAL table
Fixes #153

DUAL is a special table that is used when no other table is. For example, `SELECT 1+2` becomes `SELECT 1+2 FROM DUAL`. Because it is a special table, it does not show up in `SHOW TABLES` or anything like that. For that reason, having the analyzer to resolve that table without needing that it is on the catalog means we don't have to add special checks for this case in the logic that finds tables in the catalog.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>