sql/analyzer: implement a rule to assign indexes
Closes #176

This commit introduces the `assign_indexes` rule, which finds expressions that are indexed and pushes them down to their respective tables. This rule works in combination with the `pushdown` rule. It simply finds indexes and combines them to wrap their table node with a placeholder node that has the indexes.
Then, is in the pushdown when the table node is wrapped with the final pushed down version with the projection, filters and indexes. These two rules work together because a `sql.Indexable` table also requires the filters and the projection to be pushed down. Since this logic is already implemented in the `pushdown` rule, this avoids repeating this complicated logic by simply delegating that work to the `pushdown` rule and only finding the indexes and combining them in `assign_indexes`.

How does finding indexes work?

- Every expression is checked against the index registry to check if it's indexed.
- When expressions inside OR or AND nodes are encountered, they are combined, if such combination is possible.
  - In OR expressions, filters for the same table are combined using `Union`.
  - In AND expression, filters for the same table are combined using `Intersection`.
  - If no combination is possible, the first-most index is kept and the others are discarded.
- All indexes are grouped by its table, generating a map between table name and a `sql.IndexLookup`.

**NOTE:** multiple column indexes are not handled yet. They will be handled in a subsequent PR as a second iteration.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>