plan/index: Add tracing to index creation
Tracing is mostly done in CreateIndex so the logic does not have to be
replicated in the drivers.

Logging of rows indexed now shows the duration. It is also logged to the
span.

Error is only logged in the CreateIndex span
(plan.backgroundIndexCreate) so it is not duplicated also in the
driver's span.

Signed-off-by: Javi Fontan <jfontan@gmail.com>