sql/analyzer: add rule to evaluate filters ahead of time
Closes #280

This commit adds the `eval_filters` rule to the analyzer, which evaluates all evaluable filters in `Filter` nodes and removes any unnecessary filter. If it turns out the final result of evaluating the filters is true, independently of the non-evaluable expressions, the filter node is just removed. If it's false, the filter node is removed and replaced with a special node called `Nothing` that returns an empty result set while still holding the instance of the inner node.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>