expression: support multiple regex engines
Add support to register multiple regular expression engines. The change
comes with "native" and "oniguruma". The latter has to be compiled with
tag "oniguruma" to be registered.

Right now the session interface does not have possibility to set
configuration variables so only one is used. "oniguruma" if it is
available or native otherwise.

Signed-off-by: Javi Fontan <jfontan@gmail.com>