pilosa: use different holder per db/table pair
Pilosa holder cannot be opened more than once. It has a channel to
signal opened state (currently unused) that is closed after the first
Open. This causes panics in the next calls:

    panic: close of closed channel

The driver was using one holder for all db/tables. This called
holder.Open once per table. This works well if all indexes belong to the
same db/table but called Open several times when the indexes belong to
more than one db/table pair. Now it has a map of holders so it's only
opened once.

Signed-off-by: Javi Fontan <jfontan@gmail.com>