analyzer: fix crash with aliases used multiple times
The reorder projection rule recreates the projection with the table
columns and aliases. When an alias is used more than once in a
node expression, for example, a filter (a=1 or a=2) it is added twice.
This caused the second time to add nil as the aliases where deleted
after adding them to projections list.

Now the columns are added only once.

Signed-off-by: Javi Fontan <jfontan@gmail.com>
(cherry picked from commit 1dddaa47a8bbc6acb4f1c899057251e5cf31421f)