auth: add authentication and authorization interface
The interface has two methods:

* Mysql: returns a vitess server auth
* Allow: checks if a user has permissions needed

Currently two auth methods are implemented:

* None: allows all logins and permissions
* Native: uses mysql_native_password method to authenticate.

Native method can read users from a json file. This contains the user
credentials and permissions for the user. The credentials can be in
plain or in mysql native format. It can also contain a list of all
permissions granted to the user. If not specified it uses default
permissions that for now is "read".

    [
      {
        "name": "root",
        "password": "*9E128DA0C64A6FCCCDCFBDD0FC0A2C967C6DB36F",
        "permissions": ["read", "write"]
      },
      {
        "name": "javi",
        "password": "Passw0rd!"
      }
    ]

To enforce permissions a new validation rule can be added with the
builder:

	ab := analyzer.NewBuilder(catalog).WithAuth(userAuth)

Signed-off-by: Javi Fontan <jfontan@gmail.com>