auth: add Audit to log user interactions
Audit is an auth.Auth wrapper that does audit trail calls to log user
interaction. An AuditMethod implements functions that are called when
authentication, authorization or query is used.

    userAuth, _ := auth.NewNativeFile(userFile)
    audit := auth.NewAuditLog(logrus.New())
    userAuth = auth.NewAudit(userAuth, audit)

Audit can be wrapped by another Audit chaining several AuditMethod. Here
only AuditLog, that sends the data to logrus, is implemented.
Some changes had to be done to previous auth version and context:

* sql.Context now holds the current query
* Auth.Allowed now takes sql.Context as argument

Note: Query is still not used in this code.

Signed-off-by: Javi Fontan <jfontan@gmail.com>