server: correctly set binary charset on blob fields
Before, there was an incorrect comparison to check whether the
field was a blob or not. That resulted in all fields, including
blobs, having the UTF8 charset. Now all fields have UTF8 charset
except for BLOBs, which have binary charset.

Also, a test has been added to ensure a correct behaviour.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>