sql/analyzer: only check aliases to qualify in the topmost project
Because aliases on the topmost project or groupby were checked using
plan.InspectExpressions, it did so recursively. That was not the
intended behaviour, since it could find other aliases in a subquery
or other nodes down the tree.
Instead, now only the expressions of the project or groupby are
checked.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>