sql/plan/create_view.go: Store the original parsed definition of the view, not the analyzed version.
A problem occurs when things like the resolved table are stored in the view
definition. This changes the analyze path to analyze the view definition anew
each time, and stores the unanalyzed view definition the view registry. CREATE
VIEW still completely analyzes the view at creation time, in order to catch any
errors in the view definition.