sql/analyzer: Split analysis of subqueries so that once-after rules are applied uniformly.
This makes is so that rules through default-rules get applied as part of
analyzing the subquery. Later, early in the once-after phase, union and
subquery queries get their own once-after processing.

This allows us to keep the node tree in a state where we can apply
transformations and optimizations across Opaque node boundaries, and come back
for further processing later.