sql/plan/indexed_table_access.go: Change static index lookup nodes to not keep returning their unused key expressions.
These needs take key expressions, but they do not evaluate them. They are not
part of the evaluation tree, but they cause some problems with things like
FixFieldIndexes and prune_columns, where all GetField expressions in the plan
tree are expected to resolve to fields that are actually in scope and
resolvable at the node that is being evaluated.

This fixes a particular evaluation bug where a subquery expression in a join
condition gets moved to the filter above the joins. If the moved subquery
expression made use of a static index table lookup, the moved expression would
fail to rewrite its field indexes appropriately and the query would return
incorrect results.