sql/analyzer/validation_rules.go: Add a validation rule to assert that a subquery expression was successfully analyzed in the last pass.
The analyzer steps that analyzes subqueries allows that analysis to fail with
an error and a partial result. The idea is that it is going to happen
iteratively while the scope and table and alias names are worked out from later
analysis steps. On later iterations of the subquery analysis, the subquery will
analyze successfully.

The problem is that after the subquery has analyzed successfully, a later
transformation of plan tree (combined with bugs or unsupported features in our
engine) could cause the subquery to stop analyzing successfully.

This code change adds a validation step to the analyzer to assert that the last
run of the subquery analysis on any plan.Subquery nodes did not return any
errors, even errors that would have been tolerated by the analyzer during
default-rules iteration for example.