sql/analyzer: indexed_joins: Keep searching for usable indexes when there are NOT(? = ?) and NOT(? <=> ?) clauses in a join conjunction.
The indexed join logic is perfectly happy to go forward with a join plan if it
doesn't find a perfect index, or even a usable index at all, for one or more
table factors in the join. This logic is probably left over from when that was
not the case, but for now we make it a little more liberal to cover some cases
we need to cover for a customer.