sql/analyzer: pushdown_filters: Avoid pushing down filters below Window nodes.
If a Filter is above a Window node, the Window node will be aggregating and
computing projections based on the rows that it sees. It is not safe to push
the filter below the Window node.

The parser and analyzer currently work in concert (and must work in concert for
Window results to be correct) to keep Filter nodes which do apply to the Window
below the Window.