sql/plan: load_data: Move LoadInfile logic from server/handler to the plan node.
This puts the LoadInfile functionality as a Service on the sql.Context and has
the plan node call into it. This is a better place for it in general, rather
than inspecting the resulting plan tree in server/handler and calling adhoc
methods on the *mysql.Conn based on what we find.

Also addresses a number of (potential) issues with the existing load data local
functionality including:

* Race-y file writes and reads from a single file (.LOADDATA) for concurrent
  requests.

* Incorrect handling of scanner.Err() checking; we now correctly return the
  errors when they occur, instead of deferring them to the Close() call.

* Avoids sending LoadInfilePacket for a non-LOCAL LOAD DATA call.

The most complected part of the new implementation is the ordering constraint
around RowUpdateAccumulator iterator calling Close() on the row source before
it returns an Ok result from the Next() call. An alternative approach would be
to copy the contents to a temporary file in LoadData.RowIter().