server/handler,sql/plan/row_update_accumulator: Move the logic for generating proper OkResult values for RowsAffected and InsertId to the engine.
This makes RowUpdateAccumulator responsible for returning the right OkResult,
instead of having the server/handler hack up the values in the result after the
fact. A few enginetests needed to be updated because they asserted the wrong
values for InsertId coming back in the OkResult.

It also makes RowUpdateAccumulator responsible for setting FoundRows, RowCount
and LastInsertId. LastInsertId handling is pretty gnarly becuase we currently
don't have it coming back from the insert iterator. It should be moved there.