server/handler.go: Close the row iter in doQuery with the same context we run the query with.
If a row iter implementation captures the |ctx| from RowIter() or Next() calls,
and that |ctx| is canceled before the Close() call happens, then the
implementation may return an error on the canceled context. It is better for
our existing implementations if the Close() calls happen with the same context
as RowIter() and Next().