Fixes pushdownGroupByAliases bug with repeated projection column name.
The analyzer rule stored column names in a hash map to mirror deps on
either side of the GroupBy and Project child node. The return expression
were recreated without the table info, which is fine for most cases but
breaks when there are deps from two tables with the same column name.

This maintains table metadata for *expression.UnresolvedColumn
throughout the rule, so that the output projections maintain table
identity.