Recursive Cte's
RecursiveCte node impl with tests for joins, group by, and subquery
aliasing.

RecursiveCte's are not supported in DML statements. Nested recursive
CTE's and nested subqueries are not supported yet. Aliasing is not
neatly split down the middle on either side for RCte unions; only the
dual table can have alias clashes on either side of the CTE for now.
Cyclical recursion will timeout at depth 1000. This will likely need a
rewrite with better global aliasing to support infinitely nestable
RecursiveCte's.

Indexed table lookups are not applied to RecursiveCte tables currently
because of a bug in the memory impl.