Deduplicate IndexLookup Filters
Opportunistically eliminate filter expressions when
sql.IndexedTableAccess is used for a table scan. The original logic
respecting join scoping for applying indexes is unaffected, this simply
connects out current filter elimination logic with the expressions used
for range scan.

This requires integrators implementing sql.IndexLookup to only return
rows selected by the lookup filter.

Changes were made to ranges and datetime filters To allow integrators
to comply to this filtering standard. GMS can now distinguish between
NULL and infinity. Tests are added for IS NULL and IS NOT NULL handling
for single and multi column expressions. Engine datetimes are not truncated
to support date index lookups with the same precision as MySQL.

We currently fail to eliminate filter inversions. More work is needed
to compare inverted filters to their range scan representation.