Fix InTuple optimizer bug
Last week we fixed a dropped optimizer error return that was masking a
tuple bug. We expected a slice of interfaces, but single element tuples
return one value. We catch this in most circumstances, but missed the
multicolumn case triggered by an AND expression. The error was passing
logictests because we still return correct results even if we fail to
push a filter into a table scan.