Invalidate CachedResult on Dispose
Calling Dispose on a CachedResult would delete the underlying cache
without maintaining a flag to indicate that we had done so if we tried
to use the CachedResult again. As a result, a re-use of a Disposed
CachedResult would mistakenly think the cache was intentionally empty,
not invalidated. This distinguishes the `noCache` and empty `cacheDone`
states.