Clean up CachedResult Dispose lifecycle
Dispose should only be called on iterators whose resources are
transient. Resources that outlive a single iterator should be
Disposed at the end of the query through the Node (not iter).
A CachedResult can be safely Disposed after the reseult is
transferred to the parent HashLookup. A HashJoin should not
be disposed when its parent is an iterative node that can call
RowIter again. But if that is mistaken for some yet unforseen
edge case, disposing the CachedResult returns it to a state where
it will correctly recompute its contents rather than cycling on
a nil'd cache where |cacheDone| is left as true.