refactor(driver): store the dsn separately from the server addr
Improves the readability of the code w.r.t what purposes the Provider handles
and points to the fact that the Provider can implement ContextBuilder.

Adds a new function to access the original dsn on the Connector and Conn.

Adds a new option sql.WithInitialDatabase to set the context database.

Now both the SessionBuilder and ContextBuilder can access the DSN via either
GetSession().DSN() or Conn.DSN().

Fixes #1596

Signed-off-by: Christian Stewart <christian@paral.in>