sql/mysql_db: Rework how we store and access table data for users, grants and replica_source_info.
In general, we keep the same structure of storing the data in structs, keyed in
multimaps, and having converters for going to and from sql.Rows. We change the
following important things:

1) We more explicitly model a multimap and an indexed set of entities with
multiple keyers.

2) We add read and write locking around edits and reads of the data table.

3) We explicitly do not expose the raw indexed set or multimap from the MySQLDb
instance itself. The unprincipled access of the various *Data instances in the
old implementation was somewhat problematic.