Detect pairs of inequalities for use in RangeJoin.
Note that although this tracks whether the inequalities are open or closed, we don't do anything with that information. Turns out it doesn't matter for correctness: It's okay for the secondary row iter assumes the ranges are closed and return extra rows, because the filters still get checked in the parent.

We probably shouldn't depend on this though. I'll track the open/closeness of the ranges in the Node in a follow-up commit.