Adjust range heap coster.
In larger queries, the time spent generating and evaluating candidate rows dominates. Using seqIOCostFactor feels like a bit of a misnomer, but cpuCostFactor is too low to reflect this behavior, and this is what costInnerJoin uses.

We want to make sure that RangeHeapJoin scores worse than LookupJoins when there's a good index for lookups, but still better than loop joins. This does that.

(Also added a regression test.)