Fix information_schema row count regression (#2075)
* Fix information_schema row count regression

Ff a table does not implement RowCount() we used to use
1000 as a default value for row costing. A recent refactor changed
that to 0. This fixes information schema tables to report the 1000
value again, which is usually accurate for small databases because of default
tables and columns. I also fixed some issues with database reporting for
info schema tables.

This regression probably still exists for some dolt tables and table
functions. I will do a pass and see if I can add some more accurate
values on the Dolt side.

* make constants for default counts