processlist: Allow for killing the context associated with non-query operations like SetDB and Prepare.
The processlist maintains a Process struct for a given mysql.Conn, and
registers a context.CancelFunc for the running operation when the handler
dispatches it. This works for queries today. This PR makes it so we also
register CancelFunc callbacks for operations which touch the database but do
not put the connection into CommandQuery, such as Prepare and ComInit.