go.mod: Bump go-icu-regex. Pick up ICU4C Cgo dependency.
From this commit, go-mysql-server takes a dependency on Cgo and a dependency on ICU4C. To depend on go-mysql-server CGO must be enabled, a C++ compiler must be available and the ICU4C development libraries must be available to the C++ compiler.

On Windows, mingw with pacman -S icu is supported.

On macOS, Homebrew installed icu4c works, but only if CGO_CPPFLAGS and CGO_LDFLAGS are set to point to the installed development libraries.

On Linux, apt-get install libicu-dev, or similar.