server/handler: Add ConnectionAuthenticated callback which Vitess can call once the connection is authenticated.
Previously gms relied on the ComInitDB callback to update the processlist with
the currently authenticated user. This resulted in the processlist showing
"unauthenticated user" for connections which were already authenticated but
which had not issued a ComInitDB. Those connections could issue queries which
would also appear in the process list. Adding an explicit callback when the
authentication is successful and allowing the processlist entry to be Command
Sleep with an authenticated user even when no database is selected is more
correct behavior here.